Adding dry run option

+++ vigor/scripts/files/empty_directory_remover.py
@@ -10,7 +10,8 @@ from pathlib import Path
 
 @click.command()
 @click.argument("root_dir")
-def remove_empty_directories(root_dir: str):
+@click.option("--dry-run", is_flag=True, default=False, help="Dry Run")
+def remove_empty_directories(root_dir: str, dry_run: bool):
     """
     Finds folders in root_dir that only contain an .nfo file, without a corresponding media file.
     In this case, these folders are "empty", and should be removed.
@@ -25,10 +26,13 @@ def remove_empty_directories(root_dir: str):
             to_delete.append((files[0], path))
 
 
+    print(dry_run)
+
     for item in to_delete:
         print(f"Deleting file {item[0]} and directory {item[1]}")
-        os.remove(os.path.join(item[1], item[0]))
-        os.rmdir(item[1])
+        if not dry_run:
+            os.remove(os.path.join(item[1], item[0]))
+            os.rmdir(item[1])
 
 
 if __name__ == "__main__":
